Create global utils folder, change files to .jsx, and resolve all problems from esLint
