+++ .editorConfig
@@ -1,0 +1,5 @@
+"editor.tabSize": 4,
+
+"editor.insertSpaces": true,
+
+"editor.detectIndentation": false
