+++ src/components/Home/styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const Text = styled.h1`
+    font-size: 1.5rem;
+    text-align: center;
+    color:tomato;
+`;
+
+export const HomeWrapper = styled.section`
+    display:flex;
+    flex-direction: column;
+    width:100%;
+`;
