+++ src/components/Register/utils.js
@@ -1,12 +1,13 @@
-export const FORM_INITAL_STATE = { name: "", surname: "", login: "", password: "", email: "" };
+export const FORM_INITAL_STATE = {
+  name: '', surname: '', login: '', password: '', email: '',
+};
+
 export const FORM_CONFIG = [
   ['name', 'First name'],
   ['surname', 'Second name'],
   ['login', 'Your Login'],
   ['password', 'Your password'],
-  ['email', 'Your email']
-]
+  ['email', 'Your email'],
+];
 
-export const createFormItem = (name, placeholder, type = 'text') => {
-  return { name, placeholder, type };
-}
+export const createFormItem = (name, placeholder, type = 'text') => ({ name, placeholder, type });
